+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Note that this release introduces significant API-breaking changes compared with
   - The default form now issues a single text item with the property `prompt`.
   - Added a "persist" option to the UI configuration form to specify fields that will not be cleared after each query.
   - Multi-session support so multiple people can use it at the same time.
+  - Added ability to specify configuration values on the command line that are available to the script (also noted below)
+- Added the ability to add parameters to chatterlang_script, chatterlang_workbench, and chatterlang_serve using "--" syntax.  So "--PATH /my/path" will result in the value being added to the configuration /my/path.  For example: python -m talkpipe.app.chatterlang_script --script "INPUT FROM echo[data=$MYPATH] | listFiles | print" --MYPATH "~"
 - Updated tutorial scripts and the script names in Tutorial 3. Moved examples to the docs directory.
 - Updated the unit documentation analyzer to ignore `item` and `items` parameters. These terms are reserved for the data passed into a segment and should not be used as parameters otherwise.
 - Updated `chatterlang_workbench` to return a 413 error when a script is longer than 10,000 characters (hard-coded).
@@ -28,8 +30,6 @@ Note that this release introduces significant API-breaking changes compared with
 - Removed `accelerate` and `pypdf` as dependencies. `accelerate` was left over from an earlier version and is no longer necessary; it added unnecessary bulk.
 - When not specified, use default temperature values and pass the temperature parameter only when explicitly provided. Not all models accept a temperature.
 - Changed readtxt and readdocx to field_segments.
-- Added the ability to add parameters to chatterlang_script using "--" syntax.  So "--PATH /my/path" will result in a contant being added to the script at compile time with the value /my/path.  For example:
-python -m talkpipe.app.chatterlang_script --script "INPUT FROM echo[data=PATH] | listFiles | print" --PATH "~"
 
 ## 0.7.1
 ### New Segments and Sources